+++ models/RoleModel.go
@@ -13,7 +13,7 @@ import (
 )
 
 type Role struct {
-	Id              int       `orm:"auto"`
+	Id              int64       `orm:"auto"`
 	Remark          string      `orm:"size(120)" form:"Remark" valid:"Required;Match(/^[\u4e00-\u9fa5]+$/);MaxSize(20);MinSize(1)"`
 	//Rule            string      `orm:"size(120)" form:"Rule"`
 	Status          int         `orm:"default(2)" form:"Status" valid:"Range(1,2);"`
@@ -131,7 +131,7 @@ func GetRoleAll()(roles []orm.Params) {
 }
 
 
-func GetRoleId(roleId int) (role Role ) {
+func GetRoleId(roleId int64) (role Role ) {
 	role = Role{Id:roleId}
 	o := orm.NewOrm()
 	o.Read(&role,"Id")
